DarlingDataManagementSystem\interfaces\utility\AppBuilder: Refactored cleanUpDuplicateApps() method to only delete stored Apps whose name matches the App's AppComponentsFactory's App's name, but uniqueIds differ. This will prevent legitimate Apps from being removed.

+++ core/abstractions/utility/AppBuilder.php
@@ -164,11 +164,14 @@ abstract class AppBuilder implements AppBuilderInterface
             $appComponentsFactory->getApp()->getLocation(),
             $appComponentsFactory->getApp()->getContainer()
         ) as $component) {
-            if($component->getUniqueId() !== $appComponentsFactory->getApp()->getUniqueId()) {
+            if(
+                $component->getName() === $appComponentsFactory->getApp()->getName()
+                &&
+                $component->getUniqueId() !== $appComponentsFactory->getApp()->getUniqueId()
+            ) {
                 $appComponentsFactory->getComponentCrud()->delete($component);
             }
         }
     }
 
-#########
 }
